many: refactor snapcraft.yaml loading out of load_config
In order to have separated logical bits and clear overlaying
communicated a logical separation is introduced between the
actual snapcraft loading in raw form and its processed/validated
form.

LP: #1775849
