+++ snapcraft/internal/lifecycle/_runner.py
@@ -1,6 +1,6 @@
 # -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
 #
-# Copyright (C) 2015-2017 Canonical Ltd
+# Copyright (C) 2015-2018 Canonical Ltd
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License version 3 as
@@ -13,11 +13,13 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 import collections
 import logging
 import os
 from subprocess import check_call
 from tempfile import TemporaryDirectory
+from typing import Sequence
 
 import yaml
 from tabulate import tabulate
@@ -42,7 +44,9 @@ from ._clean import mark_dependents_dirty
 logger = logging.getLogger(__name__)
 
 
-def execute(step: steps.Step, project_options, part_names=None):
+def execute(step: steps.Step,
+            project_config: 'snapcraft.internal.project_loader._config.Config',
+            part_names: Sequence[str]=None):
     """Execute until step in the lifecycle for part_names or all parts.
 
     Lifecycle execution will happen for each step iterating over all
@@ -54,21 +58,21 @@ def execute(step: steps.Step, project_options, part_names=None):
     and after is not in this set, an exception will be raised.
 
     :param str step: A valid step in the lifecycle: pull, build, prime or snap.
-    :param project_options: Runtime options for the project.
+    :param ProjectConfig: Fully loaded project (old logic moving either to
+                          Project or the PluginHandler).
     :param list part_names: A list of parts to execute the lifecycle on.
     :raises RuntimeError: If a prerequesite of the part needs to be staged
                           and such part is not in the list of parts to iterate
                           over.
     :returns: A dict with the snap name, version, type and architectures.
     """
-    config = project_loader.load_config(project_options)
     installed_packages = repo.Repo.install_build_packages(
-        config.build_tools)
+        project_config.build_tools)
     if installed_packages is None:
         raise ValueError(
             'The repo backend is not returning the list of installed packages')
 
-    installed_snaps = repo.snaps.install_snaps(config.build_snaps)
+    installed_snaps = repo.snaps.install_snaps(project_config.build_snaps)
 
     os.makedirs(constants.SNAPCRAFT_INTERNAL_DIR, exist_ok=True)
     state_path = os.path.join(constants.SNAPCRAFT_INTERNAL_DIR, 'state')
@@ -76,21 +80,21 @@ def execute(step: steps.Step, project_options, part_names=None):
         state_file.write(yaml.dump(
             states.GlobalState(installed_packages, installed_snaps)))
 
-    if _should_get_core(config.data.get('confinement')):
-        _setup_core(project_options.deb_arch,
-                    config.data.get('base', 'core'))
+    if _should_get_core(project_config.data.get('confinement')):
+        _setup_core(project_config.project.deb_arch,
+                    project_config.data.get('base', 'core'))
 
-    executor = _Executor(config, project_options)
+    executor = _Executor(project_config)
     executor.run(step, part_names)
     if not executor.steps_were_run:
         logger.warn(
             'The requested action has already been taken. Consider\n'
             'specifying parts, or clean the steps you want to run again.')
 
-    return {'name': config.data['name'],
-            'version': config.data.get('version'),
-            'arch': config.data['architectures'],
-            'type': config.data.get('type', '')}
+    return {'name': project_config.data['name'],
+            'version': project_config.data.get('version'),
+            'arch': project_config.data['architectures'],
+            'type': project_config.data.get('type', '')}
 
 
 def _setup_core(deb_arch, base):
@@ -156,10 +160,10 @@ def _replace_in_part(part):
 
 class _Executor:
 
-    def __init__(self, config, project_options):
-        self.config = config
-        self.project_options = project_options
-        self.parts_config = config.parts
+    def __init__(self, project_config):
+        self.config = project_config
+        self.project = project_config.project
+        self.parts_config = project_config.parts
         self.steps_were_run = False
         self._steps_run = collections.defaultdict(set)
         self._dirty_reports = collections.defaultdict(dict)
@@ -330,9 +334,7 @@ class _Executor:
         if step == steps.PRIME and part_names == self.config.part_names:
             common.env = self.config.snap_env()
             meta.create_snap_packaging(
-                self.config.data, self.config.parts, self.project_options,
-                self.config.snapcraft_yaml_path,
-                self.config.original_snapcraft_yaml,
+                self.config.data, self.config.parts, self.project,
                 self.config.validator.schema)
 
     def _handle_dirty(self, part, step, dirty_report, cli_config):
